change type on randomUtil

+++ Core/Utils/RandomUtil.cs
@@ -104,7 +104,7 @@ public class RandomUtil
     /// </summary>
     /// <param name="chancePercent">The percentage chance (0-100) that the event will occur.</param>
     /// <returns>`true` if the event occurs, `false` otherwise.</returns>
-    public bool GetChance100(float chancePercent)
+    public bool GetChance100(double chancePercent)
     {
         chancePercent = Math.Clamp(chancePercent, 0f, 100f);
